ENH: Simplify and make helper scripts work on Windows
Instead of using shell scripts, which are hard to execute on Windows ported the scripts to Python.
Using Python also allowed the scripts to be greatly improved and simplified.

Also removed redundant storage of index files (in revision control, as release asset, and release note). Now csv file is only stored as release asset and md file is stored as release asset and node. No need to commit anything in revision control when files are added or removed.

Upload command uploads incoming files to all hashing algorithms. It can also creates all requested releases.

Download commands handles multiple versions of the same filename correctly (adds hash to the end of the filename if there are multiple versions).